tests/wpt/tests/css/css-text/i18n/css3-text-line-break-baspglwj-115.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Text, linebreaks: U+10A52 KHAROSHTHI PUNCTUATION CIRCLE</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://drafts.csswg.org/css-text-3/#line-breaking'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='should dom'>
<meta name="assert" content="[Exploratory] The browser will break a line of text after any Unicode character with the BA Other Terminating Punctuation property.">
<style type='text/css'>
.test > div { font-family: monospace; font-size: 25px; width: 50ch; line-height: 30px; }
#breakable2 { white-space: pre-line; }
#breakable3 { white-space: pre-wrap; }
#breakable4 { white-space: break-spaces; }
</style>
</head>
<body>



<div class="test">
  <div id="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x10A52;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
  <div id="breakable2">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x10A52;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
  <div id="breakable3">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x10A52;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
  <div id="breakable4">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x10A52;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
</div>


<!--Notes:

Box height: <span id='debugresult'></span><script>document.getElementById('debugresult').innerHTML = document.getElementById('breakable').offsetHeight;</script>

-->
<script>
test(function() {
  assert_true(document.getElementById('breakable').offsetHeight > 35);
}, "white-space:normal");
test(function() {
  assert_true(document.getElementById('breakable2').offsetHeight > 35);
}, "white-space:pre-line");
test(function() {
  assert_true(document.getElementById('breakable3').offsetHeight > 35);
}, "white-space:pre-wrap");
test(function() {
  assert_true(document.getElementById('breakable4').offsetHeight > 35);
}, "white-space:break-spaces");
</script>

<div id='log'></div>

</body>
</html>
