Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-text/i18n/css3-text-line-break-baspglwj-115-expected.txt

aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa𐩒bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa𐩒bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa𐩒bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa𐩒bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb

PASS white-space:normal
PASS white-space:pre-line
PASS white-space:pre-wrap
PASS white-space:break-spaces
